Add the CPPAD_TAPE_ID_TYPE setting to the configure command line and test using a signed integer type.
git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@2458 0e562018-b8fb-0310-a933-ecf073c7c197

+++ cppad/local/par_var.hpp
@@ -83,7 +83,7 @@ namespace CppAD {
 	bool Parameter(const AD<Base> &x)
 	{	if( x.tape_id_ == 0 )
 			return true;
-		size_t thread = x.tape_id_ % CPPAD_MAX_NUM_THREADS;
+		size_t thread = size_t(x.tape_id_ % CPPAD_MAX_NUM_THREADS);
 		return x.tape_id_ != *AD<Base>::tape_id_ptr(thread); 
 	}
 
@@ -92,7 +92,7 @@ namespace CppAD {
 	bool Parameter(const VecAD<Base> &x)
 	{	if( x.tape_id_ == 0 )
 			return true;
-		size_t thread = x.tape_id_ % CPPAD_MAX_NUM_THREADS;
+		size_t thread = size_t(x.tape_id_ % CPPAD_MAX_NUM_THREADS);
 		return x.tape_id_ != *AD<Base>::tape_id_ptr(thread); 
 	}
 
@@ -102,7 +102,7 @@ namespace CppAD {
 	bool Variable(const AD<Base> &x)
 	{	if( x.tape_id_ == 0 )
 			return false;
-		size_t thread = x.tape_id_ % CPPAD_MAX_NUM_THREADS;
+		size_t thread = size_t(x.tape_id_ % CPPAD_MAX_NUM_THREADS);
 		return x.tape_id_ == *AD<Base>::tape_id_ptr(thread); 
 	}
 
@@ -111,7 +111,7 @@ namespace CppAD {
 	bool Variable(const VecAD<Base> &x)
 	{	if( x.tape_id_ == 0 )
 			return false;
-		size_t thread = x.tape_id_ % CPPAD_MAX_NUM_THREADS;
+		size_t thread = size_t(x.tape_id_ % CPPAD_MAX_NUM_THREADS);
 		return x.tape_id_ == *AD<Base>::tape_id_ptr(thread); 
 	}
 }
